# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Torneo de Tenis",
        "version": "1.0.0"
    },
    "paths": {
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get all games",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Create a new game",
                "operationId": "b8b9557618eb0f7b3650bdd4ab3a6487",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get a game",
                "operationId": "f0bce1dcff3d56d19826d671761a43ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of game to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Update a game",
                "operationId": "7a7a51e92437ab9405a4bd9153eacec7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of game to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Delete a game",
                "operationId": "1b52414d7af18b35901add3451cba91c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of game to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get all players",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Create a new player",
                "operationId": "4e2752a0305de0adfa2a8c9ac1aa37a9",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/players/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get a player",
                "operationId": "dbef515d1b5f749e736bca18efe444e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of player to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Update a player",
                "operationId": "b0080e3337c0615ba30865afac7ad398",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of player to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Delete a player",
                "operationId": "a0d43e3c36f501bca2882e19d853c171",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of player to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/tournaments": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get all tournaments",
                "operationId": "8626d4f0c19b7c5d27dce357833b6b97",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Create a new tournament",
                "operationId": "d6053f0972883878f37e936e0316a92d",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/tournaments/{id}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Get a tournament",
                "operationId": "6fc0b01dfb2e637254d85be733c11867",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Update a tournament",
                "operationId": "8005525d118e8b5faf9780fe28777f31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Delete a tournament",
                "operationId": "a34335cdf270ec1cbbe62c0062364dc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}